Development

+++ pkg/api/client/service.go
@@ -14,7 +14,7 @@ func NewServiceClient(c ScoretrakClient) service.Serv {
 	return &ServiceClient{c}
 }
 
-func (s ServiceClient) Delete(id uint64) error {
+func (s ServiceClient) Delete(id uint32) error {
 	return s.s.GenericDelete(fmt.Sprintf("/service/%d", id))
 }
 
@@ -27,7 +27,7 @@ func (s ServiceClient) GetAll() ([]*service.Service, error) {
 	return sg, nil
 }
 
-func (s ServiceClient) GetByID(id uint64) (*service.Service, error) {
+func (s ServiceClient) GetByID(id uint32) (*service.Service, error) {
 	sg := &service.Service{}
 	err := s.s.GenericGet(sg, fmt.Sprintf("/service/%d", id))
 	if err != nil {
@@ -44,7 +44,7 @@ func (s ServiceClient) Update(u *service.Service) error {
 	return s.s.GenericUpdate(u, fmt.Sprintf("/service/%d", u.ID))
 }
 
-func (s ServiceClient) TestService(id uint64) (*queueing.ScoringData, error) {
+func (s ServiceClient) TestService(id uint32) (*queueing.ScoringData, error) {
 	sd := &queueing.ScoringData{}
 	err := s.s.GenericGet(sd, fmt.Sprintf("/service/test/%d", id))
 	if err != nil {
