+++ pkg/storage/orm/service_repo.go
@@ -17,7 +17,7 @@ func NewServiceRepo(db *gorm.DB, log logger.LogInfoFormat) service.Repo {
 	return &serviceRepo{db, log}
 }
 
-func (s *serviceRepo) Delete(id uint64) error {
+func (s *serviceRepo) Delete(id uint32) error {
 	s.log.Debugf("deleting the service with id : %d", id)
 
 	result := s.db.Delete(&service.Service{}, "id = ?", id)
@@ -46,7 +46,7 @@ func (s *serviceRepo) GetAll() ([]*service.Service, error) {
 	return services, nil
 }
 
-func (s *serviceRepo) GetByID(id uint64) (*service.Service, error) {
+func (s *serviceRepo) GetByID(id uint32) (*service.Service, error) {
 	s.log.Debugf("get service details by id : %s", id)
 
 	ser := &service.Service{}
